backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
firebase-admin==6.3.0
pandas==2.1.3
plaid-python==21.1.0
httpx==0.25.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
cryptography==41.0.7
aioredis==2.0.1
celery==5.3.4
redis==5.0.1
alembic==1.12.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.34.0
tenacity==8.2.3
bcrypt==4.0.1
